various alterations to allow the speed of light to be user-set.  Up to now, it uses c=1 by default; but Ungar suggests that sometimes one might like to investigate what happens when c tends to infinity.
The method is to use something like options(c=299792458).

The various functions that use the speed of light, viz as.3vel(),
r3vel(), and add3(), use the same mechanism as in the onion package.

It needs more documentation, in particular, it is pretty inadvisable
to change the speed of light half way through a calculation.